Refactoring hdf5widget as an example
